+++ Tests/BugsnagPerformanceTests/BugsnagPerformanceConfigurationTests.mm
@@ -8,6 +8,7 @@
 
 #import <XCTest/XCTest.h>
 #import <BugsnagPerformance/BugsnagPerformance.h>
+#import "BugsnagPerformanceConfiguration+Private.h"
 
 @interface BugsnagPerformanceConfigurationTests : XCTestCase
 
@@ -49,6 +50,25 @@
     NSError *error = nil;
     XCTAssertFalse([config validate:&error]);
     XCTAssertNotNil(error);
+}
+
+- (void)testShouldSendReportsWhenEnabledReleaseStagesAreEmpty {
+    auto config = [[BugsnagPerformanceConfiguration alloc] initWithApiKey:@"0123456789abcdef0123456789abcdef"];
+    XCTAssertTrue([config shouldSendReports]);
+}
+
+- (void)testShouldSendReportsWhenEnabledReleaseStagesContainCurrentStage {
+    auto config = [[BugsnagPerformanceConfiguration alloc] initWithApiKey:@"0123456789abcdef0123456789abcdef"];
+    config.enabledReleaseStages = [NSSet setWithArray: @[@"env1", @"env2"]];
+    config.releaseStage = @"env2";
+    XCTAssertTrue([config shouldSendReports]);
+}
+
+- (void)testShouldNotSendReportsWhenEnabledReleaseStagesDontContainCurrentStage {
+    auto config = [[BugsnagPerformanceConfiguration alloc] initWithApiKey:@"0123456789abcdef0123456789abcdef"];
+    config.enabledReleaseStages = [NSSet setWithArray: @[@"env1", @"env2"]];
+    config.releaseStage = @"env3";
+    XCTAssertFalse([config shouldSendReports]);
 }
 
 @end